updating the file configuration for PPO algorithm renewal
