+++ Battery-Control-By-Reinforcement-Learning/sample.py
@@ -1,0 +1,10 @@
+def plus(x, y):
+  print(f"{x} + {y} =")
+  return x + y
+
+a = "Hello"
+b = " World!"
+c = plus(a, b)
+print(c)
+c = plus(2, 3)
+print(c)
